+++ Axys/Extensions/Entities/ModelEntity+Sphere.swift
@@ -9,8 +9,11 @@ import RealityKit
 import UIKit
 
 extension ModelEntity {
-	static func movableSphere() -> ModelEntity {
-		let sphere = ModelEntity(mesh: .generateSphere(radius: 0.01), materials: [UnlitMaterial(color: .white)])
+	static func markerSphere() -> ModelEntity {
+		let sphere = ModelEntity(
+			mesh: .generateSphere(radius: 0.01),
+			materials: [UnlitMaterial(color: .white)]
+		)
 		return sphere
 	}
 }
