+++ Axys/Extensions/Logger+Ext.swift
@@ -13,5 +13,6 @@ extension Logger {
 	static let calibration = Logger(subsystem: subsystem, category: "Calibration")
 	static let connection = Logger(subsystem: subsystem, category: "Connection")
     static let rhino = Logger(subsystem: subsystem, category: "Rhino")
+	static let models = Logger(subsystem: subsystem, category: "Models")
     static let views = Logger(subsystem: subsystem, category: "Views")
 }
