+++ Axys/Models/Axis/Axes.swift
@@ -27,29 +27,3 @@ enum Axes: CaseIterable, Identifiable {
         }
     }
 }
-
-enum SegmentedMode {
-    case position, rotation
-}
-
-struct AxisOptions: OptionSet {
-    let rawValue: Int
-
-    static let x = AxisOptions(rawValue: 1 << 0)
-    static let y = AxisOptions(rawValue: 1 << 1)
-    static let z = AxisOptions(rawValue: 1 << 2)
-
-    static let xy: AxisOptions = [.x, .y]
-    static let all: AxisOptions = [.x, .y, .z]
-}
-
-
-extension AxisOptions {
-    mutating func toggle(_ axis: AxisOptions) {
-        if contains(axis) {
-            remove(axis)
-        } else {
-            insert(axis)
-        }
-    }
-}
