+++ Axys/Models/Axis/SegmentedMode.swift
@@ -1,0 +1,10 @@
+//
+//  SegmentedMode.swift
+//  Axys
+//
+//  Created by Alessandro Bortoluzzi on 04/06/25.
+//
+
+enum SegmentedMode {
+    case position, rotation
+}
